backend/responses/serializers.py
from rest_framework import serializers

from responses.models import Answer, Response


class AnswerSerializer(serializers.ModelSerializer):
    class Meta:
        model = Answer
        fields = ('id', 'value', 'response')


class ResponseSerializer(serializers.ModelSerializer):
    class Meta:
        model = Response
        fields = ('id', 'status')
